+++ Flow.Launcher.Plugin.VisualStudio/Images/delete.png
