+++ Flow.Launcher.Plugin.VisualStudio/Images/folder.png
