+++ Flow.Launcher.Plugin.VisualStudio/Images/icon.png
