+++ Flow.Launcher.Plugin.VisualStudio/Images/notification.png
