+++ Flow.Launcher.Plugin.VisualStudio/Images/windows.png
