+++ Flow.Launcher.Plugin.VisualStudio/Main.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Flow.Launcher.Plugin.VisualStudio.Models;
 
 namespace Flow.Launcher.Plugin.VisualStudio
 {
