+++ Flow.Launcher.Plugin.VisualStudio/UI/SettingsView.xaml.cs
