+++ Flow.Launcher.Plugin.VisualStudio/UI/Styles.xaml
