+++ Flow.Launcher.Plugin.VisualStudio/UI/VisualStudioViewModel.cs
@@ -1,6 +1,6 @@
 namespace Flow.Launcher.Plugin.VisualStudio.UI;
 
-public class VisualStudioModel
+public class VisualStudioViewModel
 {
 	public string IconPath { get; init; }
 	public string Name { get; init; }
