+++ Flow.Launcher.Plugin.VisualStudio/plugin.json
@@ -4,7 +4,7 @@
     "Name": "Visual Studio Launcher",
     "Description": "Open your recent solutions, projects and files in Visual Studio",
     "Author": "Odotocodot",
-    "Version": "1.0.4",
+    "Version": "1.1.0-beta",
     "Language": "csharp",
     "Website": "https://github.com/Odotocodot/VisualStudio4Flow",
     "IcoPath": "Images\\icon.png",
