--- appveyor.yml
@@ -1,15 +1,0 @@
-version: 1.0.0.{build}
-image: Visual Studio 2022
-
-environment:
-    DOTNET_SKIP_FIRST_TIME_EXPERIENCE: true
-    DOTNET_CLI_TELEMETRY_OPTOUT: 1
-
-build_script:
-- ps: dotnet publish -c Release -r win-x64 --no-self-contained Flow.Launcher.Plugin.VisualStudio/Flow.Launcher.Plugin.VisualStudio.csproj
-
-after_build:
-- ps: Compress-Archive -Path "Flow.Launcher.Plugin.VisualStudio\bin\Release\win-x64\publish\*" -DestinationPath "Flow.Launcher.Plugin.VisualStudio.zip"
-
-artifacts:
-    - path: 'Flow.Launcher.Plugin.VisualStudio.zip'
